make local test request between services

+++ deployment/initial-db-sql/01-schema.sql
@@ -1,12 +1,12 @@
 CREATE TABLE products (
-    id BIGINT PRIMARY KEY,
+    id serial PRIMARY KEY,
     name VARCHAR(255),
     price_per_piece DECIMAL(19,2),
     stock_level BIGINT
 );
 
 CREATE TABLE orders (
-    id BIGINT PRIMARY KEY,
+    id serial PRIMARY KEY,
     customer_id BIGINT,
     status VARCHAR(100)
 );
